refactor line of code for cve to each own class
