+++ lib/salus.rb
@@ -3,6 +3,7 @@ require 'active_support/core_ext'
 require 'salus/bugsnag'
 require 'salus/cli'
 require 'salus/repo'
+require 'salus/package_utils'
 require 'salus/scanners'
 require 'salus/config'
 require 'salus/config_exception'
